Inclusão de buscar por id, delete por id e salvar e alteração do repository pra Long

+++ src/main/java/com/redelog/api/repository/EntregaRepository.java
@@ -3,5 +3,5 @@ package com.redelog.api.repository;
 import com.redelog.api.model.entities.Entrega;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EntregaRepository extends JpaRepository<Entrega, Integer> {
+public interface EntregaRepository extends JpaRepository<Entrega, Long> {
 }
